# apps/services/src/services/openapi-specs/openapi.sql.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sql2api Sqls API",
    "description": "SQL 注册与管理模块：类型判定、静态审计 + AI 审查/生成、CRUD。注册成功后可通过 /openapi/invoke/{id} 调用。禁止 DROP / DELETE / TRUNCATE。多语句或 CALL 归为 complex（POST）。所有业务接口需 Bearer Api-Key 鉴权。",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:13334",
      "description": "Local development"
    }
  ],
  "tags": [
    {
      "name": "Sqls",
      "description": "SQL 语句注册、AI 审查/生成与 CRUD"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/openapi/sqls": {
      "post": {
        "tags": [
          "Sqls"
        ],
        "summary": "注册 SQL",
        "description": "解析 SQL 类型并映射 HTTP method（SELECT→GET、INSERT→POST、UPDATE→PATCH、complex→POST）。多语句或 CALL/函数归为 complex。静态审计硬性禁止 DROP/DELETE/TRUNCATE，再经 AI 审查语法与性能；未通过返回 422。",
        "operationId": "createSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSqlBody"
              },
              "example": {
                "connection_id": "550e8400-e29b-41d4-a716-446655440000",
                "name": "get-user-by-id",
                "description": "按 id 查询用户",
                "sql": "SELECT id, name, email FROM users WHERE id = :id",
                "params": [
                  {
                    "name": "id",
                    "rule": "required|integer",
                    "description": "用户 ID"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SqlItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "description": "AI 审查未通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFailedResponse"
                },
                "example": {
                  "request_id": "req_xxx",
                  "timestamp": 1710000000000,
                  "code": "422",
                  "message": "SQL Review Failed",
                  "data": {
                    "passed": false,
                    "issues": [
                      {
                        "severity": "error",
                        "message": "DELETE without WHERE clause",
                        "suggestion": "Add a WHERE condition to limit affected rows"
                      }
                    ]
                  }
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "get": {
        "tags": [
          "Sqls"
        ],
        "summary": "SQL 列表",
        "operationId": "listSqls",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "按名称模糊搜索"
          },
          {
            "name": "connection_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sql_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "select",
                "insert",
                "update",
                "complex"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SqlListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/openapi/sqls/generate": {
      "post": {
        "tags": [
          "Sqls"
        ],
        "summary": "AI 生成 SQL",
        "description": "根据自然语言与连接下的模型上下文生成 SQL，不落库。",
        "operationId": "generateSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSqlBody"
              },
              "example": {
                "connection_id": "550e8400-e29b-41d4-a716-446655440000",
                "prompt": "查询指定用户的全部订单",
                "model_ids": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResultResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/openapi/sqls/review": {
      "post": {
        "tags": [
          "Sqls"
        ],
        "summary": "AI 审查 SQL",
        "description": "静态审计（禁止 DROP/DELETE/TRUNCATE）+ AI 审查语法/性能。可选传入 connection_id 以加载方言与模型上下文；不落库。响应额外返回推断的 sql_type 与 method。",
        "operationId": "reviewSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSqlBody"
              },
              "example": {
                "sql": "SELECT id, name FROM users WHERE id = :id",
                "connection_id": "550e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "审查完成（无论 passed 与否均 200）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResultResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/openapi/sqls/{id}/openapi": {
      "get": {
        "tags": [
          "Sqls"
        ],
        "summary": "获取单条 SQL 的 OpenAPI 文档",
        "description": "返回自包含的迷你 OpenAPI 3.0 文档（含该 SQL 的 invoke path、参数 schema、共享响应组件），便于复制到 ApiFox 等工具。disabled 的 SQL 也可获取。",
        "operationId": "getSqlOpenApiDoc",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SqlId"
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 文档对象（位于响应 envelope 的 data 字段）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeBase"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "description": "OpenAPI 3.0.3 document",
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/openapi/sqls/{id}": {
      "get": {
        "tags": [
          "Sqls"
        ],
        "summary": "SQL 详情",
        "operationId": "getSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SqlId"
          }
        ],
        "responses": {
          "200": {
            "description": "详情成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SqlItemResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Sqls"
        ],
        "summary": "更新 SQL",
        "description": "若变更 sql 文本，将重新判定类型并 AI 审查。",
        "operationId": "updateSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SqlId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSqlBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SqlItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "description": "AI 审查未通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewFailedResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Sqls"
        ],
        "summary": "删除 SQL",
        "operationId": "deleteSql",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SqlId"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ApiKey",
        "description": "Authorization: Bearer <api-key>。通过 `sql2api apikey create` 生成，明文仅展示一次。"
      }
    },
    "parameters": {
      "SqlId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "已注册 SQL 的 UUID"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未鉴权或 Api-Key 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "同名 SQL 已存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "AI 模型不可用",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiEnvelopeBase": {
        "type": "object",
        "required": [
          "request_id",
          "timestamp",
          "code",
          "message",
          "data"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
      "ErrorEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        ]
      },
      "SqlParamDef": {
        "type": "object",
        "required": [
          "name",
          "rule"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "与 SQL 中 :name 占位符对应"
          },
          "rule": {
            "type": "string",
            "description": "validatorjs 规则，如 required|integer"
          },
          "description": {
            "type": "string"
          },
          "default": {
            "description": "缺省时使用的默认值"
          }
        }
      },
      "ReviewIssue": {
        "type": "object",
        "required": [
          "severity",
          "message"
        ],
        "properties": {
          "severity": {
            "type": "string",
            "enum": [
              "error",
              "warning",
              "info"
            ]
          },
          "message": {
            "type": "string"
          },
          "suggestion": {
            "type": "string"
          }
        }
      },
      "ReviewResult": {
        "type": "object",
        "required": [
          "passed",
          "issues"
        ],
        "properties": {
          "passed": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewIssue"
            }
          },
          "sql_type": {
            "type": "string",
            "enum": [
              "select",
              "insert",
              "update",
              "complex"
            ],
            "description": "审查端点返回时附带的推断类型"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PATCH"
            ],
            "description": "审查端点返回时附带的推断 HTTP method"
          }
        }
      },
      "CreateSqlBody": {
        "type": "object",
        "required": [
          "connection_id",
          "name",
          "sql"
        ],
        "properties": {
          "connection_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SqlParamDef"
            }
          }
        }
      },
      "UpdateSqlBody": {
        "type": "object",
        "properties": {
          "connection_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SqlParamDef"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ]
          }
        }
      },
      "GenerateSqlBody": {
        "type": "object",
        "required": [
          "connection_id",
          "prompt"
        ],
        "properties": {
          "connection_id": {
            "type": "string",
            "format": "uuid"
          },
          "prompt": {
            "type": "string"
          },
          "model_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "可选，过滤参与上下文的模型 id"
          }
        }
      },
      "ReviewSqlBody": {
        "type": "object",
        "required": [
          "sql"
        ],
        "properties": {
          "sql": {
            "type": "string"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SqlItem": {
        "type": "object",
        "required": [
          "id",
          "app_id",
          "connection_id",
          "name",
          "description",
          "sql",
          "sql_type",
          "method",
          "endpoint",
          "params",
          "status",
          "review",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "app_id": {
            "type": "string",
            "format": "uuid"
          },
          "connection_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "sql": {
            "type": "string"
          },
          "sql_type": {
            "type": "string",
            "enum": [
              "select",
              "insert",
              "update",
              "complex"
            ]
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PATCH"
            ]
          },
          "endpoint": {
            "type": "string",
            "description": "调用路径，形如 /openapi/invoke/{id}"
          },
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SqlParamDef"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "disabled"
            ]
          },
          "review": {
            "$ref": "#/components/schemas/ReviewResult"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GenerateResult": {
        "type": "object",
        "required": [
          "sql",
          "sql_type",
          "method",
          "params",
          "explanation"
        ],
        "properties": {
          "sql": {
            "type": "string"
          },
          "sql_type": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SqlParamDef"
            }
          },
          "explanation": {
            "type": "string"
          }
        }
      },
      "SqlListData": {
        "type": "object",
        "required": [
          "list",
          "total",
          "page",
          "size"
        ],
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SqlItem"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "DeletedData": {
        "type": "object",
        "required": [
          "id",
          "deleted"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        }
      },
      "SqlItemResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SqlItem"
              }
            }
          }
        ]
      },
      "SqlListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SqlListData"
              }
            }
          }
        ]
      },
      "GenerateResultResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/GenerateResult"
              }
            }
          }
        ]
      },
      "ReviewResultResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ReviewResult"
              }
            }
          }
        ]
      },
      "ReviewFailedResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "422"
              },
              "message": {
                "type": "string",
                "example": "SQL Review Failed"
              },
              "data": {
                "$ref": "#/components/schemas/ReviewResult"
              }
            }
          }
        ]
      },
      "DeletedResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiEnvelopeBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DeletedData"
              }
            }
          }
        ]
      }
    }
  }
}